3208 update the plasma profiles section of the docs to include all new models (#3256)
* Legacy initial commit

* Initial commit

* chore: Refactor Profile class and add documentation

Refactor the Profile class in profiles.py to improve code structure and readability. Also, add detailed documentation to explain the purpose and functionality of the class and its methods.

* refactor: Update Plasma Profiles navigation structure

Update the navigation structure in the mkdocs.yml file to include sub-pages for the Profiles section. This change adds separate pages for the Overview, Density Profile, Temperature Profile, and Profile Base Class. It also updates the Composition & Impurities page and adds a new Radiation page. This improves the organization and accessibility of the Plasma Profiles documentation.

* Legacy initial commit

* chore: Refactor Plasma Profiles documentation structure and add Profile class documentation

* Temperature profile page first commit

* refactor: Improve Profile class structure and documentation

Refactor the Profile class in profiles.py to enhance code structure and readability. Additionally, add comprehensive documentation to clarify the purpose and functionality of the class and its methods.

* refactor: Improve Profile class structure and documentation

* refactor: Update Plasma Profiles parameterization method

Update the parameterization method in the PlasmaProfile class to handle the case where ipedestal is equal to 0. This improves the accuracy and reliability of the plasma profile calculations.

* Added discussion about current profile, constraint equations and derivative plots for the profile graidants

* refactor: Update Plasma Profiles parameterization method

* Replace tcore calculation with new equation

* refactor: Update plasma temperature profile calculation method

* Remove old plasma_profiles.md

* Add switch for core desnity calc depending on ipedestal. Tidy up loose docstring

* Spelling and grammar corrections

* Corrected typos

* Tidied up maths and corrected typos

* Remove ieped and eped_sf

* Fix profiles.py docstring grammar

* Remove ieped and eped_sf from test files

---------

Co-authored-by: Jack Foster <jack.foster@ukaea.uk>