954 bootstrap current fraction scaling formula for steady state machines (#3268)
* Added new bootstap model switch

* feat: Add Sakai bootstrap current fraction model

This commit adds a new bootstrap current fraction model based on the Sakai formula. The `bootstrap_fraction_sakai` method is implemented in the `Physics` class in the `physics.py` file. This model calculates the bootstrap fraction using a set of parameters such as plasma beta, bootstrap current coefficient, safety factor, neoclassical viscosity parameters, and epsilon. The formula is derived from the work of Sakai et al. [^1^].

[^1^]: Sakai, R., Fujita, T., Okamoto, A. (2019). Derivation of bootstrap current fraction scaling formula for 0-D system code analysis. Fusion Engineering and Design, 149, 111322. https://doi.org/10.1016/j.fusengdes.2019.111322

* feat: Implement Sakai bootstrap current fraction model

* Update docstring

* feat: Add Sakai bootstrap current fraction model docs

* Added unit test for large tokamak and ST case

* feat: Update Sakai bootstrap current calculation

This commit modifies the `Physics` class in the `physics.py` file to update the calculation of the Sakai bootstrap current. The Sakai formula states that the ACCOME dataset used has the toroidal diamagnetic current included in the bootstrap current. Therefore, the commit removes the PROCESS calculated diamagnetic current from the bootstrap current calculation. This change improves the accuracy of the bootstrap current fraction model.

* Added level 3 warning about calculating diamagnetic current when using the Sakai scaling